python-api/requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]

# Data Validation and Configuration
pydantic
pydantic-settings
python-multipart

# Supabase (PostgreSQL and S3)
supabase  # Official Python client for Supabase

# LLM and File Processing (Placeholders for now)
google-genai
openai

# Environment Management
python-dotenv

# docling document processing
docling
docling-core
Pillow
ImageHash
